refactor: update file name to match function
Relates #166
