+++ supabase/modals/getItems.ts
@@ -1,4 +1,4 @@
-import newClient from '../utils/createClient';
+import newClient from '../utils/newClient';
 import { PartialItem } from '../../types/types';
 
 export async function getItems(
